add kafka events

+++ planner-core-api/src/main/kotlin/com/kvsinyuk/plannercoreapi/model/request/CreateProductRequestDto.kt
@@ -1,8 +1,10 @@
 package com.planner.dto.request
 
+import java.math.BigDecimal
+
 data class CreateProductRequestDto(
     val categoryId: Long,
     val userId: Long,
-    val price: Float,
+    val price: BigDecimal,
     val description: String = ""
 )
